+++ vaultaire/src/main/kotlin/com/github/manosbatsis/vaultaire/dsl/ConditionsDsl.kt
@@ -55,9 +55,9 @@ interface Condition {
 }
 
 /** A [Condition] that contains other conditions. Allows for nested and/or condition groups */
-abstract class ConditionsCondition<T : StatePersistable, out F: Fields<T>>: Condition {
+abstract class ConditionsCondition<P : StatePersistable, out F: Fields<P>>: Condition {
 
-    /** The fields of the target [StatePersistable] type `T` */
+    /** The fields of the target [StatePersistable] type `P` */
     abstract val fields: F
 
     val conditions: MutableList<Condition> = mutableListOf()
@@ -66,9 +66,9 @@ abstract class ConditionsCondition<T : StatePersistable, out F: Fields<T>>: Cond
         conditions += condition
     }
 
-    fun and(initializer: CompositeCondition<T, F>.() -> Unit) = addCondition(AndCondition(fields).apply(initializer))
+    fun and(initializer: CompositeCondition<P, F>.() -> Unit) = addCondition(AndCondition(fields).apply(initializer))
 
-    fun or(initializer: CompositeCondition<T, F>.() -> Unit) = addCondition(OrCondition(fields).apply(initializer))
+    fun or(initializer: CompositeCondition<P, F>.() -> Unit) = addCondition(OrCondition(fields).apply(initializer))
 
 }
 
@@ -76,59 +76,53 @@ abstract class ConditionsCondition<T : StatePersistable, out F: Fields<T>>: Cond
  * A [ConditionsCondition], base implementation for and/or condition group containers.
  *  Allows for query criteria conditions.
  */
-abstract class CompositeCondition<T : StatePersistable, out F: Fields<T>>(
+abstract class CompositeCondition<P : StatePersistable, out F: Fields<P>>(
         override val fields: F
-) : ConditionsCondition<T, F>() {
+) : ConditionsCondition<P, F>() {
 
-    fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.isNull() =
+    fun <S : Comparable<S>> FieldWrapper<P, S>.isNull() =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.isNull())))
 
-    @supress("unused")
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.isNull(none: Unit) = isNull()
-
-    fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.notNull() =
+    fun <S : Comparable<S>> FieldWrapper<P, S>.notNull() =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notNull())))
 
-    @supress("unused")
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.notNull(none: Unit) = notNull()
-
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.equal(value: S) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.equal(value: S) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.equal(value))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.`==`(value: S) = equal(value)
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.`==`(value: S) = equal(value)
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.notEqual(value: S) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.notEqual(value: S) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notEqual(value))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.`!=`(value: S) = notEqual(value)
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.`!=`(value: S) = notEqual(value)
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.lessThan(value: S) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.lessThan(value: S) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.lessThan(value))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.lt(value: S) = lessThan(value)
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.lt(value: S) = lessThan(value)
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.lessThanOrEqual(value: S) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.lessThanOrEqual(value: S) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.lessThanOrEqual(value))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.lte(value: S) = lessThanOrEqual(value)
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.lte(value: S) = lessThanOrEqual(value)
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.greaterThan(value: S) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.greaterThan(value: S) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.greaterThan(value))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.gt(value: S) = greaterThan(value)
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.gt(value: S) = greaterThan(value)
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.greaterThanOrEqual(value: S) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.greaterThanOrEqual(value: S) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.greaterThanOrEqual(value))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.gte(value: S) = greaterThanOrEqual(value)
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.gte(value: S) = greaterThanOrEqual(value)
 
-    fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.between(from: S, to: S) =
+    fun <S : Comparable<S>> FieldWrapper<P, S>.between(from: S, to: S) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.between(from, to))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.between(value: Pair<S, S>) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.between(value: Pair<S, S>) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.between(value.first, value.second))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.btw(value: Pair<S, S>) = between(value)
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.btw(value: Pair<S, S>) = between(value)
 
     infix fun <T : StatePersistable> FieldWrapper<T, String>.like(value: String) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.like(value))))
@@ -142,37 +136,37 @@ abstract class CompositeCondition<T : StatePersistable, out F: Fields<T>>(
     fun <T : StatePersistable> FieldWrapper<T, String>.notLike(value: String, exactMatch: Boolean = true) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notLike(value, exactMatch))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.isIn(value: Collection<S>) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.isIn(value: Collection<S>) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.`in`(value))))
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.`in`(value: Collection<S>) = isIn(value)
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.`in`(value: Collection<S>) = isIn(value)
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.notIn(value: Collection<S>) =
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.notIn(value: Collection<S>) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notIn(value))))
 
-    fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.`!in`(value: Collection<S>) = notIn(value)
+    fun <S : Comparable<S>> FieldWrapper<P, S>.`!in`(value: Collection<S>) = notIn(value)
 
     // non type safe versions
-    infix fun <T : StatePersistable> FieldWrapper<T, *>._equal(value: Any) =
+    infix fun FieldWrapper<P, *>._equal(value: Any) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.equal(value))))
 
-    infix fun <T : StatePersistable> FieldWrapper<T, *>._notEqual(value: Any) =
+    infix fun FieldWrapper<P, *>._notEqual(value: Any) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.property.notEqual(value))))
 
-    infix fun <T : StatePersistable> FieldWrapper<T, *>._like(value: String) =
+    infix fun FieldWrapper<P, *>._like(value: String) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.asStringProperty().like(value))))
 
-    infix fun <T : StatePersistable> FieldWrapper<T, *>._notLike(value: String) =
+    infix fun FieldWrapper<P, *>._notLike(value: String) =
             addCondition(VaultCustomQueryCriteriaCondition(VaultCustomQueryCriteria(this.asStringProperty().notLike(value))))
 
     @supress("UNCHECKED_CAST")
-    protected fun <T : StatePersistable> FieldWrapper<T, *>.asStringProperty() = this.property as KProperty1<T, String>
+    protected fun FieldWrapper<P, *>.asStringProperty() = this.property as KProperty1<P, String>
 }
 
 /** Defines a set of conditions where all items must be matched */
-class AndCondition<T : StatePersistable, out F: Fields<T>>(
+class AndCondition<P : StatePersistable, out F: Fields<P>>(
         fields: F
-) : CompositeCondition<T, F>(fields) {
+) : CompositeCondition<P, F>(fields) {
 
     override fun toCriteria(): QueryCriteria? {
         var criteria: QueryCriteria? = null
@@ -188,9 +182,9 @@ class AndCondition<T : StatePersistable, out F: Fields<T>>(
 }
 
 /** Defines a set of conditions where at least a single item must be matched */
-class OrCondition<T : StatePersistable, out F: Fields<T>>(
+class OrCondition<P : StatePersistable, out F: Fields<P>>(
         fields: F
-) : CompositeCondition<T, F>(fields) {
+) : CompositeCondition<P, F>(fields) {
 
     override fun toCriteria(): QueryCriteria? {
         var criteria: QueryCriteria? = null
@@ -215,24 +209,22 @@ class VaultCustomQueryCriteriaCondition(val criterion: VaultCustomQueryCriteria<
 }
 
 /** Used to define [Sort.SortColumn]s */
-class SortColumns<T : StatePersistable>(val statePersistableType: Class<T>){
+class SortColumns<P : StatePersistable>(val statePersistableType: Class<P>){
 
     val ASC = Sort.Direction.ASC
     val DESC = Sort.Direction.DESC
     val entries: LinkedHashSet<Sort.SortColumn> = linkedSetOf()
 
-    infix fun <T : StatePersistable, S : Comparable<S>> FieldWrapper<T, S>.sort(value: Sort.Direction) =
-            entries.add(Sort.SortColumn(SortAttribute.Custom(statePersistableType, this.property.name), value))
-
+    infix fun <S : Comparable<S>> FieldWrapper<P, S>.sort(value: Sort.Direction) {
+        entries.add(Sort.SortColumn(SortAttribute.Custom(statePersistableType, this.property.name), value))
+    }
 }
-
-
 
 /**
  * A [ConditionsCondition] extended by Vaultaire's annotation processing to create a condition DSL specific to a [ContractState] type.
  * Defines a root [QueryCriteria.VaultQueryCriteria]. Allows for defining `and`/`or` condition groups, as well as a [Sort]
  */
-abstract class VaultQueryCriteriaCondition<T : StatePersistable, out F: Fields<T>>(
+abstract class VaultQueryCriteriaCondition<P : StatePersistable, out F: Fields<P>>(
         var status: Vault.StateStatus = Vault.StateStatus.UNCONSUMED,
         var stateRefs: List<StateRef>? = null,
         var notary: List<AbstractParty>? = null,
@@ -242,17 +234,17 @@ abstract class VaultQueryCriteriaCondition<T : StatePersistable, out F: Fields<T
         var constraintTypes: Set<Vault.ConstraintInfo.Type> = emptySet(),
         var constraints: Set<Vault.ConstraintInfo> = emptySet(),
         var participants: List<AbstractParty>? = null
-) : ConditionsCondition<T, F>() {
+) : ConditionsCondition<P, F>() {
 
     /** The target [ContractState] type */
     abstract val contractStateType: Class<out ContractState>
 
     /** The target [StatePersistable] type */
-    abstract val statePersistableType: Class<T>
+    abstract val statePersistableType: Class<P>
 
-    lateinit var sortColumns: SortColumns<T>
+    lateinit var sortColumns: SortColumns<P>
 
-    fun orderBy(initializer: SortColumns<T>.() -> Unit) {
+    fun orderBy(initializer: SortColumns<P>.() -> Unit) {
         sortColumns = SortColumns(statePersistableType).apply(initializer)
     }