+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and can be helpful with version upgrades.
 ## 0.2 
 
 - Renamed `VaultQueryDsl` annotation to `VaultaireGenerate`
-- Added `FieldsAwareStateService`. The annotation processor will now 
-generate a subclass per annotated `PersistentState`
+- Renamed `StateService` to `BasicStateService`, added extended `StateService` type.
+- The annotation processor will now generate a subclass of `StateService` per annotated `PersistentState`
 
 ## 0.1